# perf(DX): Reduce recorder overhead