Recorder has performance penalty while running in prod. This PR attempts to reduce it a little by:
- deferring `EXPLAIN` of queries
- deferring duplicate counts
- deferring query formatting with sqlparse


No scientific way to measure it but general but overhead should go down by 20-60% depending upon usage. 

| action | without recorder | w/ Recorder before | w/ recorder now | reduction in overhead |
| --- | --- | ---| ---| --- |
| loading `/` | 160ms | 300 ms | 225ms | ~50% |